acla_ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.7
python-multipart==0.0.6
httpx==0.25.2
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
asyncio==3.4.3
aiofiles==23.2.1
